Super-terse ipfs downloader added to util.  Pass it a cid and a path, make download happen.  No configuraton.
