+++ main.go
@@ -3,19 +3,13 @@ package main
 import (
 	"context"
 	"fmt"
-	"io"
-	"net/url"
 	"os"
 	"os/signal"
-	gopath "path"
-	"path/filepath"
+	"path/filepath"	
 	"strings"
 	"syscall"
 
-	"github.com/cheggaaa/pb"
-	files "github.com/ipfs/go-ipfs-files"
 	iface "github.com/ipfs/interface-go-ipfs-core"
-	ipath "github.com/ipfs/interface-go-ipfs-core/path"
 	"github.com/ipfs/ipget/get"
 	cli "github.com/urfave/cli/v2"
 )
@@ -61,7 +55,7 @@ func main() {
 		}
 
 		outPath := c.String("output")
-		iPath, err := parsePath(c.Args().First())
+		iPath, err := get.ParsePath(c.Args().First())
 		if err != nil {
 			return err
 		}
@@ -103,7 +97,7 @@ func main() {
 			}
 			return cli.Exit(err, 2)
 		}
-		err = WriteTo(out, outPath, c.Bool("progress"))
+		err = get.WriteTo(out, outPath, c.Bool("progress"))
 		if err != nil {
 			if err == context.Canceled {
 				return <-sigExitCoder
@@ -158,80 +152,4 @@ func movePostfixOptions(args []string) []string {
 
 	// append extracted arguments to the real args
 	return append(args, endArgs...)
-}
-
-func parsePath(path string) (ipath.Path, error) {
-	ipfsPath := ipath.New(path)
-	if ipfsPath.IsValid() == nil {
-		return ipfsPath, nil
-	}
-
-	u, err := url.Parse(path)
-	if err != nil {
-		return nil, fmt.Errorf("%q could not be parsed: %s", path, err)
-	}
-
-	switch proto := u.Scheme; proto {
-	case "ipfs", "ipld", "ipns":
-		ipfsPath = ipath.New(gopath.Join("/", proto, u.Host, u.Path))
-	case "http", "https":
-		ipfsPath = ipath.New(u.Path)
-	default:
-		return nil, fmt.Errorf("%q is not recognized as an IPFS path", path)
-	}
-	return ipfsPath, ipfsPath.IsValid()
-}
-
-// WriteTo writes the given node to the local filesystem at fpath.
-func WriteTo(nd files.Node, fpath string, progress bool) error {
-	s, err := nd.Size()
-	if err != nil {
-		return err
-	}
-
-	var bar *pb.ProgressBar
-	if progress {
-		bar = pb.New64(s).Start()
-	}
-
-	return writeToRec(nd, fpath, bar)
-}
-
-func writeToRec(nd files.Node, fpath string, bar *pb.ProgressBar) error {
-	switch nd := nd.(type) {
-	case *files.Symlink:
-		return os.Symlink(nd.Target, fpath)
-	case files.File:
-		f, err := os.Create(fpath)
-		defer f.Close()
-		if err != nil {
-			return err
-		}
-
-		var r io.Reader = nd
-		if bar != nil {
-			r = bar.NewProxyReader(r)
-		}
-		_, err = io.Copy(f, r)
-		if err != nil {
-			return err
-		}
-		return nil
-	case files.Directory:
-		err := os.Mkdir(fpath, 0777)
-		if err != nil {
-			return err
-		}
-
-		entries := nd.Entries()
-		for entries.Next() {
-			child := filepath.Join(fpath, entries.Name())
-			if err := writeToRec(entries.Node(), child, bar); err != nil {
-				return err
-			}
-		}
-		return entries.Err()
-	default:
-		return fmt.Errorf("file type %T at %q is not supported", nd, fpath)
-	}
 }
